mutex: Fix missing check in shared_mutex
shared_mutex has a while loop in the unlock function, which only checked if
the first waiter in the list had a desired state of shared.
If that condition is true, we pop that waiter and loop again.
This can lead to a nullptr access if we pop the entire list.
Adding an extra check for an empty list fixes this issue.
Bug found by members of the OSDev discord while debugging Chromium.

Signed-off-by: Dennis Bonke <admin@dennisbonke.com>