wait-group: implement Go-ish wait groups
One of the few niceties of Go is it's WaitGroup abstraction. It is
essentially a semaphore that starts on a predetermined number, and
counts down on zero, and wakes up any waiters (and any future waiters).
Most of the code for this class matches oneshot events (since a oneshot
event is a non-rearmable one-member wait group), and as such, a followup
will reimplement oneshot_event in terms of wait_groups (and
coincidentally fix a minor bug where cancellation status was not
correctly reported when a wait was cancelled).